% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_to_df.R
\name{a_to_df}
\alias{a_to_df}
\title{Funkcja a_to_df}
\usage{
a_to_df(x, save = TRUE)
}
\arguments{
\item{x}{ścieżka do pliku .mat}

\item{save}{TRUE/FALSE określający czy obiekt wynikowy ma być zapisany do pliku. W przypadku opcji TRUE, obiekt zapisywany jest do tego samego miejsca na dysku, gdzie znajduje się plik .mat.}
}
\value{
obiekt klasy data.frame lub plik .csv
}
\description{
Ta funkcja zamienia wielowymiarową macierz na obiekt klasy data.frame lub na plik .csv.
}
\examples{
\donttest{
   plik <- 'C:/Users/kot/dane/10symulacji.mat'
   a_to_df(plik, save=TRUE)
}
}

