% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_changer.R
\name{text_changer}
\alias{text_changer}
\title{Funkcja text_changer}
\usage{
text_changer(file, t1, t2, t3)
}
\arguments{
\item{file}{ścieżka do pliku RDC}

\item{t1}{tekst do zastąpienia wartości w wierszach 16 i 18}

\item{t2}{tekst do zastąpienia wartości w wierszu 22}

\item{t3}{tekst do zastąpienia wartości w wierszu 22}
}
\value{
nowy plik ze zmienionymi wartościami.
Nowy plik będzie miał taką samą nazwę z dopiskiem _new
}
\description{
Ta funkcja zamienia wartości parametrów w wierszach 16,
18, 22, 23 w plikach o rozszerzeniu RDC.
}

