% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_changer.R
\name{mt_changer}
\alias{mt_changer}
\title{Funkcja mt_changer}
\usage{
mt_changer(folder_path, t1, t2, t3, ...)
}
\arguments{
\item{folder_path}{ścieżka do folderu zawierającego pliki RDC}

\item{t1}{tekst do zastąpienia wartości w wierszach 16 i 18}

\item{t2}{tekst do zastąpienia wartości w wierszu 22}

\item{t3}{tekst do zastąpienia wartości w wierszu 22}
}
\value{
nowe pliki ze zmienionymi wartościami.
Nowy plik będzie miał taką samą nazwę w folderze new
}
\description{
Ta funkcja zamienia wartości parametrów w wierszach 16, 18, 22, 23
w folderze zawierającym pliki o rozszerzeniu RDC.
}
\examples{
\donttest{
   folder <- "C:/Users/jakub/pliki_RDC"
   mt_changer(folder, "7500.0000000", "0.0000000", "Background")
}
}

